Move extensions to the base classes
And add XML comments to all public types/members
