+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/Channels/ClosedChannelException.cs
@@ -7,10 +7,16 @@ namespace System.Threading.Tasks.Channels
     /// <summary>Exception thrown when a channel is used after it's been closed.</summary>
     public class ClosedChannelException : InvalidOperationException
     {
+        /// <summary>Initializes a new instance of the <see cref="ClosedChannelException"/> class.</summary>
         public ClosedChannelException() : base(Properties.Resources.ClosedChannelException_DefaultMessage) { }
 
+        /// <summary>Initializes a new instance of the <see cref="ClosedChannelException"/> class.</summary>
+        /// <param name="message">The message that describes the error.</param>
         public ClosedChannelException(string message) : base(message) { }
 
+        /// <summary>Initializes a new instance of the <see cref="ClosedChannelException"/> class.</summary>
+        /// <param name="message">The message that describes the error.</param>
+        /// <param name="innerException">The exception that is the cause of this exception.</param>
         public ClosedChannelException(string message, Exception innerException) : base(message, innerException) { } 
     }
 }
