+++ src/System.Threading.Tasks.Channels/System/Threading/Tasks/VoidResult.cs
@@ -4,5 +4,6 @@
 
 namespace System.Threading.Tasks
 {
+    /// <summary>An empty struct, used to represent void in generic types.</summary>
     internal struct VoidResult { }
 }
